requirements.txt
flask==3.0.0
flask-cors==4.0.0
replicate==0.25.1
requests==2.31.0
gunicorn==21.2.0
Pillow==10.2.0
